+++ hardhat.config.js
@@ -1,5 +1,5 @@
-const merge = require('lodash.merge');
+const {mergeConfigs} = require('@animoca/ethereum-contracts/src/config');
 
 require('@animoca/ethereum-contracts/hardhat-plugins');
 
-module.exports = merge(require('@animoca/ethereum-contracts/hardhat-config'), require('./hardhat-config'));
+module.exports = mergeConfigs(require('@animoca/ethereum-contracts/hardhat-config'), require('./hardhat-config'));